requirements.txt
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.14.0
greenlet==3.0.1
psycopg2==2.9.9
pytz==2023.3.post1
SQLAlchemy==2.0.23
sqlparse==0.4.4
typing_extensions==4.8.0
